refactor: show orig extrinsic
