+++ src/renderer/entities/transaction/lib/transactionService.ts
@@ -1,14 +1,16 @@
 import { type ApiPromise } from '@polkadot/api';
 import { type SubmittableExtrinsic } from '@polkadot/api/types';
 import { type SignerOptions } from '@polkadot/api/types/submittable';
-import { GenericSignerPayload } from '@polkadot/types';
+import { Compact, Enum, GenericCall, GenericMultiAddress, GenericSignerPayload } from '@polkadot/types';
 import { type ExtrinsicEra, type Weight } from '@polkadot/types/interfaces';
+import { type Codec } from '@polkadot/types/types';
 import { hexToU8a, u8aToHex } from '@polkadot/util';
 import { blake2AsU8a, signatureVerify } from '@polkadot/util-crypto';
 import { merkleizeMetadata } from '@polkadot-api/merkleize-metadata';
 
 import {
   type Address,
+  type Chain,
   type HexString,
   type MultisigAccount,
   type MultisigTxWrapper,
@@ -19,7 +21,9 @@ import {
   type Wallet,
   WrapperKind,
 } from '@/shared/core';
+import { toAddress } from '@/shared/lib/utils';
 import { type TxMetadata, createTxMetadata, dictionary, nullable, upgradeNonce } from '@/shared/lib/utils';
+import { pjsSchema } from '@/shared/polkadotjs-schemas';
 import { type AccountId } from '@/shared/polkadotjs-schemas';
 // TODO transaction service should be inside network domain
 // eslint-disable-next-line boundaries/element-types
@@ -48,6 +52,8 @@ export const transactionService = {
   getExtrinsicWeight,
   getTxWeight,
   verifySignature,
+
+  formatExtrinsic,
 
   logPayload,
 };
@@ -361,6 +367,76 @@ function verifySignature(payload: Uint8Array, signature: HexString, accountId: A
   } catch {
     return false;
   }
+}
+
+function formatArg(arg: Codec, chain: Chain): unknown {
+  if (Array.isArray(arg)) {
+    return arg.map((a) => formatArg(a, chain));
+  }
+
+  if (arg instanceof GenericCall) {
+    const args: Record<string, unknown> = {};
+
+    for (const [key, value] of arg.argsEntries) {
+      args[key] = formatArg(value, chain);
+    }
+
+    return {
+      section: arg.section,
+      method: arg.method,
+      args,
+    };
+  }
+
+  if (arg instanceof GenericMultiAddress) {
+    if (arg.type === 'Id' || arg.type === 'Address20' || arg.type === 'Address32') {
+      return toAddress(arg.value.toString(), { prefix: chain.addressPrefix });
+    } else {
+      return arg.toHuman();
+    }
+  }
+
+  const isAccount = pjsSchema.accountId.safeParse(arg);
+  if (isAccount.success) {
+    return toAddress(isAccount.data, { prefix: chain.addressPrefix });
+  }
+
+  if (arg instanceof Compact) {
+    return formatArg(arg.unwrap(), chain);
+  }
+
+  if (arg instanceof Enum) {
+    return {
+      [arg.type]: formatArg(arg.value, chain),
+    };
+  }
+
+  if (arg instanceof Map) {
+    const result: Record<string, unknown> = {};
+
+    for (const [key, value] of arg) {
+      result[key] = formatArg(value, chain);
+    }
+
+    return result;
+  }
+
+  return arg.toHuman();
+}
+
+function formatExtrinsic(extrinsic: Extrinsic, chain: Chain): object {
+  const args: Record<string, unknown> = {};
+
+  // @ts-expect-error argsEntries are not defined in extrinsic type
+  for (const [key, value] of extrinsic.method.argsEntries as [string, Codec][]) {
+    args[key] = formatArg(value, chain);
+  }
+
+  return {
+    section: extrinsic.method.section,
+    method: extrinsic.method.method,
+    args,
+  };
 }
 
 function logPayload(info: Awaited<ReturnType<typeof createPayloadWithMetadata | typeof createPayloadWithProof>>[]) {
